MB-62036: [BP] Add cm_rebalance_in_progress
This backports:
MB-61813: Add cm_rebalance_in_progress

The cm_rebalance_in_progress stat is a gauge returning whether a
rebalance is running (1) or not running (0). This is only reported on
the orchestrator node in order to be consistent with other rebalance
progress stats.

Change-Id: Ief9f51ccc145d76749106af01ecada4f1459db62
Reviewed-on: https://review.couchbase.org/c/ns_server/+/210474
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>