MB-63070: [BP] don't attempt to redact stats
This backports https://review.couchbase.org/c/ns_server/+/213760,
incorporating the one line fix-up in
https://review.couchbase.org/c/ns_server/+/214618.
This is needed because the fix for MB-63070 makes it much more likely
for redacting stats to cause a cbcollect failure.

Original commit message:

We were accidentally attempting to redact data in the binary stats data.
This wasn't causing any problems because it was never matching, but it
was still incorrectly attempting to redact a lot of data it didn't need
to. Also rename buffer variable because of pylint warnings.

Change-Id: I2eafdea7f2e1696c22e9f72360565c49abd3653d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/214893
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker