MB-62900: Disable AFO in node_remap script
We are seeing auto failovers on turn on of Capella clusters as some
nodes take longer to turn on than others. The solution to this issue
is to turn off AFO before turning a cluster on, but that is not an
appropriate solution for snapshot backup and restore (to a different
cluster) as we would be turning off and on auto-failover potentially
every hour.

Add a capability to disable auto-failover to the node remap script.

Change-Id: Ifbf5c60299571bdd5e54fd9bea300e6e1bf218d0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/213606
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>