MB-62900: Add support to only disable AFO via config_remap
Such that we can disable AFO on an offline cluster, a requirement of
snapshot and backup to the same cluster as we need to disable AFO
before restoring the cluster (while the nodes are offline) but do
not want to remap them. This is necessary as we are seeing spurious
auto failovers in Capella clusters on turn on/restore as all nodes are
brought up at about the same time and some may take longer to come up
than others

This requires that we support running config_remap without remap args.

Change-Id: Ied12d4c47b5423471a038d0f9bf31b0630513978
Reviewed-on: https://review.couchbase.org/c/ns_server/+/213656
Well-Formed: Restriction Checker
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>