MB-63312: Remove old assertion + fix passthrough
This fixes two issues. Firstly it removes the assertion that used to
enforce that the very end of the file must contain a newline.

The second is actually more problematic since it was causing us to not
collect stats data at all on redacted versions of the zipfiles. With
these changes, there is only 1 file different from the
redacted/unredacted now which is user.dets which is completely skipped
for the redacted one.

Before this, there were many missing files/folders related to the
prometheus stats.

Change-Id: I63f7d5722f08048352c3db76e58cf39b7153e086
Reviewed-on: https://review.couchbase.org/c/ns_server/+/215070
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker