MB-63891: Simplify good/bad test handling in bucket tests
We now determine whether a test case is expected to fail or not based
on the get_errors function. This means we no longer need to make sure
that the test cases in init_test_params will all be valid requests,
which greatly simplifies updating the tests.

This change also removes all the use of good_count and bad_count, since
these don't make sense with this change.

The testset can be simplified quite substantially more, and needs a lot
more comments to make it make more sense, but that work will be done in
Morpheus.

Change-Id: I81b26cb6a4fc465b6da9859cdcd5bca43eec32c4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/218060
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>