MB-63942 - Show index path when fts present but not gsi
GSI, FTS, and View indexes use the same storage path setting in
ns_server. The path is referred to as the "Index Storage Path".

The original bug noted that this path was not being shown on nodes
where the FTS service is present, but the GSI (index) service is not
present.

However, it was later noted that since the path is used for Views as
well, the "Index" path should be shown any time there is a node with
KV, Index, or FTS.

The fix involves a few tweaks:

- Don't show the "index" path as part of the normal iteration of
  service path information.

- Instead, have an HTML block showing the "index" path if any of "kv",
  "index", or "fts" is present.

Change-Id: I6a95aee163fd0bb8d980b635016ec31d90f3c14a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/218244
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>