Add chronicle_compat_events supports to fake_chronicle_kv
So that we can easily test config changes via chronicle_compat_events.
This capability is not immediately useful, but it does remove a usage
of fake_ns_pubsub which lets us remove it in the future.

Moving the ns_pubsub mock narrows down the usages of it.

Change-Id: I68495dd86799f5a032631d6a18ef87dd70e8ad7b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/230912
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>