MB-67857[BP]: Refactor to avoid fake_ns_pubsub uses
The following cherry-picks only work when merged together, because of
a weird bug with meck and ns_pubsub which causes a failure after the
first commit, but is resolved by the second commit.

Original commit messages as follows

-----------------------------------------------------------------------

Set up the leader_registry properly in failover_tests

We should set up leader_registry properly in failover_tests (i.e.
by using the gen_event for leader_events) such that we don't have
to rely on the fake ns_pubsub mechanism.

This is part of a larger change to move away from fake_ns_pubsub which
adds complexity that is not necessary as setting up these gen_events
is simple.

Change-Id: I750b13256cbd56fb63a0c4630cdfe2fd28200613

-----------------------------------------------------------------------

Add chronicle_compat_events supports to fake_chronicle_kv

So that we can easily test config changes via chronicle_compat_events.
This capability is not immediately useful, but it does remove a usage
of fake_ns_pubsub which lets us remove it in the future.

Moving the ns_pubsub mock narrows down the usages of it.

Change-Id: I68495dd86799f5a032631d6a18ef87dd70e8ad7b

Change-Id: Ib318f0c3c109d41dbbc9a8c97cefb69d471b0a18
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232056
Well-Formed: Restriction Checker
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>