[BP] Set up the leader_registry properly in failover_tests
We should set up leader_registry properly in failover_tests (i.e.
by using the gen_event for leader_events) such that we don't have
to rely on the fake ns_pubsub mechanism.

This is part of a larger change to move away from fake_ns_pubsub which
adds complexity that is not necessary as setting up these gen_events
is simple.

Change-Id: I750b13256cbd56fb63a0c4630cdfe2fd28200613
Reviewed-on: https://review.couchbase.org/c/ns_server/+/230908
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>