MB-66105: [BP] fix loading default profile
This patch does a few things to simplify/improve the profile loading
code. Firstly, we use the CB_FORCE_PROFILE mechanism in cluster_run as
we do for all of the other profiles. That way we won't mistakenly read
from the /etc/couchbase.d/config_profile signal file that's used in
production. That said, I also made the constraints much tighter,
requiring a successful reading of the particular profile file or we
crash. Originally when the default profile was empty, it made more sense
to allow this generic "fallback" behavior. Now that the default profile
contains keys, this likely isn't "good enough" and instead acts as a
sort of development "foot gun".

Now we MUST feed in a correct config_path, and we must be able to load
the *_profile file or we will crash the server. This should help in
catching issues in this area quicker.

We will return 'default' if no other name is found and we accept any
{error, _} for this situation instead of just enoent. Particularly,
eacces is also commonly seen as an error type.

Also needed to add code to couchdb because since we operate it as a
separate "node" it needs to populate it's persistent_term database
individually. This means config_path is required for any "nodes" where
they will use config_profile keys.

Change-Id: I75a8759a117f38ba0ea1b69776b8c881d85981f5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226250
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker