MB-65785: Add workload_pattern_default bucket config
To address a behaviour change in memcached, we must add a new bucket
configuration parameter in a maintenance release such that the user
can access the old behaviour in a persistent, upgradeable, manner.

This parameter should not be advertised in any documentation til
Morpheus, at which point we know that the cluster will fully support it.
In a patch against Morpheus, we will ugprade the bucket configuration
to always store this bucket configuration parameter. Til then, the
REST API will return `undefined` from nodes that know about this
parameter, until the user sets it to some value.

Change-Id: Idb977ec6b4dac58680d0736659a7b4b388851fb9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226846
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>