MB-66415: Add test tags for filtering cluster_tests
Adds a tag (decorator) which can be applied to tests which we wish to
mark as belonging to some category.

By default, all tests will not have any tags, which will mean that
calling cluster_tests with --with-tags will not run any tests, and
--without-tags <tag>(,...) will run all tests.

In order to have validation of tags for local use, while at the same
time being able to specify tags in CV jobs across multiple branches
which may not all recognise all the tags, we also have the
--ignore-unknown-tags flag, which is primarily intended for use by
CV jobs.

Change-Id: Ic75cab9e9ddcc021ca8a12ed6185d5537374834f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226832
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Restriction Checker