Generate vbmap in failover tests
Such that we can rely on more correct maps in tests in the future.

This change particularly improves the stale config tests by generating a
map for the active KV nodes whilst taking into account the inactive ones
(but only those that are inactiveFailed) such that the map more
accurately reflects one that would be seen post-failover.

Change-Id: I9f04efb0df324d44bf88cced611c39de486bcb0c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227348
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>