MB-66720: [BP] Prevent external user admin access to local user
This is a backport of: MB-65113

Prior to this change a user with the security_admin_external role
could obtain info for a local user. This is now prevented.

The backported change is slightly different than the newer releasess in
the ordering of the "domain" and "exists" checks. This will be resolved
when this change is merged forward.

Change-Id: I848904a553325abb06175654be9a97ce1e1d0a1d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227804
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>