Deep sort rebalance_details tests
A curious issue occurs when adding the graceful failover unit test. The
ordering of elements created when we create a map during the
rebalance_details call changes. As a map is an unordered structure this
is valid, but we are converting it to a list and comparing it to another
in this test. Lists being ordered caused this test to fail. Handle this
by sorting the lists.

It's not clear why the ordering changes if we run the graceful failover
test first.

Change-Id: I8a07042e8d8eb0239cf19188a2e38f5a5ca4f7b2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227894
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>