MB-66730: create_bucket should include collection manifest
This allows memcached to avoid having to use a cache for the manifest,
which has complexity with encrypting that file for encryption at rest.

We provide the manifest as base64url-encoded, to avoid any future risks
of "="/";" ending up in the manifest and breaking the bucket config
parsing (without escaping those characters).

Since the collection manifest can be very large, we don't want to dump
the whole manifest as part of the bucket config, so instead we're just
logging the uid.

Prior to this change, the collection manifest was being set almost
immediately after creating the bucket. That logic is not being removed,
since it covers the ensure_bucket case when the bucket already exists.

Change-Id: Iac9df3317d1f1c14664b14d550f069267fad13cd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228107
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>