cluster_tests: Add num_vbuckets whenever a bucket is used
Previously we implicitly had 16 vbuckets for all tests, but this can't
be relied upon, since a testset is being added with 1024 vbuckets,
which now means that other testsets may run with 1024 vbuckets.

Since the 16 vbucket default was for test efficiency purposes, not just
arbitrary, we should make this explicit on all applicable tests.

Change-Id: Ie40cdca4a54aec9d4a0de75fbb6f8b2b4c776ab0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228396
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>