MB-66679: Provide auto-failover settings in profile
The "provisioned_profile" now contains the default auto-failover
settings to use instead of those in ns_server. This provides the ability
for capella to provision clusters with their customizations without
having to apply them via REST.

On an upgrade only newly defined key/values in the profile are added
to the auto-failover config. Other, existing key/values in the profile
are ignored.

The auto-failover settings in the serverless profile now match those
in the provisioned profile. It's felt those would be more in line with
the needs should serverless become viable again.

Change-Id: I1acf6158f6098eec50a2746577e9fe6299fed004
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228765
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>