MB-67045:[BP] Override to disable clustermap push
Due to a bug in libcouchbase we need an option to disable the ability
to push clustermap notifications from memcached. By default memcached
will always allow the client to enable all requested features it
supports, but in this case it needs a setting to allow the system to
refuse enable of clustermap push.

We are adding this config key to memcached.json via the following API:

POST /pools/default/settings/memcached/global
 -d clustermap_push_notifications_enabled=false

While we would typically wish to gate something like this by cluster
compat mode, in this case we are relying on this feature not being
documented until 8.0, and not including any value by default (using the
memcached_config_extra key).

Change-Id: I7caa4667b1286b16e122aeaf4225c85dd72cc03b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229468
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>