MB-55506: Log error info when ensure_directories fails
Previously when path_config:ensure_directories failed it would simply
fail with {badmatch, error}. This wasn't providing enough information to
triage the issue.

With this change we now log the directory path, the error that occurred,
and what ns_server expects (a directory that is creatable/writable).

Change-Id: I2b26dd33c22687204429b8c45513b596f5b9174c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228819
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>