MB-61292: Fix native encryption tests test_teardown
In case if some test fails and leaves log encryption enabled
it is possible that test_teardown will fail to remove the secret
that encrypts the logs. In order to work around this issue,
test_teardown changes encryption to nodeSecretManager, which
reencrypts the logs deks with master password, and only after
that disables the encryption.

Change-Id: Iac506205bc01bf636f4724583f88faf0879085bd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228877
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>