Use specific map in auto failover snapshot test
The randomly generated map sometimes means that we trip over earlier
auto-failover checks that we are not trying to test here. Use a
specific map such that we will pass the original auto-failover checks
and then later trip over the failover checks after we sync the config.

Change-Id: I857c7ec6d9d8aa9b64285a1e69cb1da873ec8f76

Reviewed-on: https://review.couchbase.org/c/ns_server/+/229095
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>