MB-67106: Deprecate kv_ep_cache_miss_ratio
The kv_ep_cache_miss_ratio stat doesn't appear to be documented and
it's name is misleading. The stat is not a cache miss ratio but a miss
rate: misses/gets. The miss could be due to a invalid key so isn't
really a cache miss.

Change-Id: I4bb9c6c05f2e83f3b3c9b79199411b105a70f5be
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229385
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>