MB-61292: Don't reread good DEKs...
... and try reusing the keys that are already in snapshot instead.

If we have an error reading the keys, we will retry reading them
after a short delay.
When we do that, there is no point in re-reading the keys that
were already read successfully.
This is especially important when there are many DEKs and slow
external KMS is used to encrypt them. Each attempt to re-read all
keys may take a long time.

Change-Id: Ie11ce67acc2928d4bb3d084f46318115aef49340
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229450
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>