MB-65045: Allow "Query List Index" role to list indexes
The permissions for the query_list_index role needs "read" permission
instead of "list" permission in order to use /indexStatus. This is needed
to list the index definitions. Note adding "read" permission does not
allow reading the index data. In fact, query doesn't use "read" on index
so there's no impact to query with the change.

Change-Id: I007e2233905fd6a832165db9a951c478aeaaa8a3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229724
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>