MB-65092 KMIP Passphrase fixes.
KMIP passphrase should be empty in Add Encryption Key dialog
until the user types something into the field.
The browser should not autocomplete the field with some password,
so autocomplete="new-password" attribute is used.

In Edit Encryption Key dialog the passphrase should be completed
if there is a passphrase set; a passphrase is set if the API
returns ****** in "keyPassphrase" property.

Change-Id: I2bcd165c6525062cb07294139e03f139d9a9fad9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/230275
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>