MB-67699: [cbcollect] add .exe on 'cygwin' ...
... as well. It was recently changed to use os.name, but that produces
the result 'posix' even on windows and this doesn't work for us in the
case of adding (or not) the .exe to binary names. Even on cygwin, the
binaries will have the .exe file extension and we would be missing them
if ran in this environment. Not sure how supported this is but I could
imagine someone ssh'ing into a windows box and attempting to run
cbcollect, which would actually run it in cygwin, and could cause
problems unless this is fixed.

Change-Id: If2b6d3d07a2e1729d8692e392970c2083ae0dedb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/230972
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>